requirements.txt
numpy==1.18.0
opencv_python==4.2.0.34
matplotlib==3.1.2
Pillow==7.1.2
